# openapi.yaml
openapi: 3.0.3
info:
  title: brain-server API
  description: >
    HTTP API for the brain-server local agent memory / semantic search server
    (Axum + rusqlite + sqlite-vec + model2vec). This spec is hand-written from
    the source in src/main.rs and src/handlers/*. Endpoints whose contract could
    not be fully verified from source are marked with `x-status: unverified`.
    Every response carries `X-Api-Version: <semver>`; legacy routes carry a
    `Deprecation` header per the deprecation policy in API_CONTRACT.md.
  version: 0.9.8
  x-api-version: "0.9.8"
servers:
  - url: http://127.0.0.1:8765
    description: Default local server (override BIND_HOST/BIND_PORT)
paths:
...
  /sources/reconcile:
    post:
      summary: Retire active sources whose live URI is no longer present
      operationId: reconcileSources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, live_uris]
              properties:
                kind: { type: string, description: e.g. "vault" }
                live_uris:
                  type: array
                  items: { type: string }
                  maxItems: 50000
      responses:
        "200":
          description: Reconcile result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted_sources: { type: integer }
                  deleted_chunks: { type: integer }
                  orphan_uris:
                    type: array
                    items: { type: string }
  /sources/{id}:
    delete:
      summary: Retire a single source by id
      operationId: deleteSource
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Source retired
        "404":
          description: No such source
...
  /webhooks/{kind}:
    post:
      summary: Verified webhook receiver (v0.9.7 Guard)
      operationId: receiveWebhook
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
          description: Connector kind, e.g. 'github'.
      responses:
        "200":
          description: Webhook accepted/enqueued (or duplicate)
        "401":
          description: Signature verification failed
...